+++ test/Microsoft.AspNetCore.SystemWebAdapters.Tests/Caching/CacheTests.cs
@@ -2,15 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections;
-using System.Collections.Generic;
 using System.IO;
 using System.Runtime.Caching;
 using System.Threading.Tasks;
-using System.Web.Hosting;
 using AutoFixture;
-using Microsoft.AspNetCore.SystemWebAdapters;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
 using Moq;
 using Xunit;
 
@@ -316,154 +312,53 @@ public class CacheTests
     }
 
     [Fact]
-    public void InsertWithDependency()
+    public void CacheFromHttpContext()
     {
         // Arrange
-        var memCache = new Mock<MemoryCache>(_fixture.Create<string>(), null);
-        var cache = new Cache(memCache.Object);
-        var key = _fixture.Create<string>();
-        var item = new object();
-        var cacheDependency = new Mock<CacheDependency>();
+        var cache = new Cache();
+
+        var serviceProvider = new Mock<IServiceProvider>();
+        serviceProvider.Setup(s => s.GetService(typeof(Cache))).Returns(cache);
+
+        var coreContext = new Mock<HttpContextCore>();
+        coreContext.Setup(c => c.RequestServices).Returns(serviceProvider.Object);
+
+        var context = new HttpContext(coreContext.Object);
 
         // Act
-        cache.Insert(key, item, cacheDependency.Object);
+        var result = context.Cache;
 
         // Assert
-        memCache.Verify(m => m.Set(key, item, It.Is<CacheItemPolicy>(e => e.AbsoluteExpiration.DateTime.Equals(Cache.NoAbsoluteExpiration) && e.SlidingExpiration.Equals(Cache.NoSlidingExpiration)), null), Times.Once);
+        Assert.Same(cache, result);
+
+        //Act via GetService
+        var cacheFromService = context.GetService<Cache>();
+        Assert.Same(cache, cacheFromService);
     }
 
     [Fact]
-    public async Task DependentFileCallback()
+    public void CacheFromHttpContextWrapper()
     {
         // Arrange
-        using var memCache = new MemoryCache(_fixture.Create<string>());
-        var cache = new Cache(memCache);
-        var item = new object();
-        var key = _fixture.Create<string>();
-        var updated = false;
-        var slidingExpiration = TimeSpan.FromMilliseconds(1);
-        CacheItemUpdateReason? updateReason = default;
+        var cache = new Cache();
 
-        var tcs = new TaskCompletionSource();
+        var serviceProvider = new Mock<IServiceProvider>();
+        serviceProvider.Setup(s => s.GetService(typeof(Cache))).Returns(cache);
 
-        void Callback(string key, CacheItemUpdateReason reason, out object? expensiveObject, out CacheDependency? dependency, out DateTime absoluteExpiration, out TimeSpan slidingExpiration)
-        {
-            expensiveObject = null;
-            dependency = null;
-            absoluteExpiration = Cache.NoAbsoluteExpiration;
-            slidingExpiration = TimeSpan.FromMilliseconds(5);
+        var coreContext = new Mock<HttpContextCore>();
+        coreContext.Setup(c => c.RequestServices).Returns(serviceProvider.Object);
 
-            updated = true;
-            updateReason = reason;
-
-            tcs.SetResult();
-        }
-
-        var file = Path.GetTempFileName();
-        await File.WriteAllTextAsync(file, key);
-
-        using var cd = new CacheDependency(file);
-        // Act
-        cache.Insert(key, item, cd, Cache.NoAbsoluteExpiration, Cache.NoSlidingExpiration, Callback);
-
-        // Ensure file is updated
-        await File.WriteAllTextAsync(file, DateTime.UtcNow.ToString("O"));
-
-        // Wait for callback to be called
-        await tcs.Task;
-
-        // Wait for callback to be finalized
-        await Task.Delay(100);
-
-        // Assert
-        Assert.True(updated);
-        Assert.Null(cache[key]);
-        Assert.Equal(CacheItemUpdateReason.DependencyChanged, updateReason);
-    }
-
-    [Fact]
-    public async Task DependentItemCallback()
-    {
-        // Arrange
-        using var memCache = new MemoryCache(_fixture.Create<string>());
-
-        var cache = new Cache(memCache);
-
-        using var _ = HostingEnvironmentContext.Initialize(cache);
-
-        var item1 = new object();
-        var item2 = new object();
-        var key1 = _fixture.Create<string>();
-        var key2 = _fixture.Create<string>();
-        var updateReason = new Dictionary<string, CacheItemUpdateReason>();
-        var slidingExpiration = TimeSpan.FromMilliseconds(1);
-
-        void Callback(string key, CacheItemUpdateReason reason, out object? expensiveObject, out CacheDependency? dependency, out DateTime absoluteExpiration, out TimeSpan slidingExpiration)
-        {
-            expensiveObject = null;
-            dependency = null;
-            absoluteExpiration = Cache.NoAbsoluteExpiration;
-            slidingExpiration = Cache.NoSlidingExpiration;
-
-            updateReason[key] = reason;
-        }
+        var context = new HttpContext(coreContext.Object);
+        var contextWrapper = new HttpContextWrapper(context);
 
         // Act
-        cache.Insert(key1, item1, null, Cache.NoAbsoluteExpiration, slidingExpiration, Callback);
-
-        using var cd = new CacheDependency(null, new[] { key1 });
-        cache.Insert(key2, item2, cd, Cache.NoAbsoluteExpiration, Cache.NoSlidingExpiration, Callback);
-
-        Assert.Empty(updateReason);
-
-        // Ensure sliding expiration has hit
-        await Task.Delay(slidingExpiration);
-
-        // Force cleanup to initiate callbacks on current thread
-        memCache.Trim(100);
+        var result = contextWrapper.Cache;
 
         // Assert
-        Assert.Contains(key1, updateReason.Keys);
-        Assert.Contains(key2, updateReason.Keys);
+        Assert.Same(cache, result);
 
-        Assert.Null(cache[key1]);
-        Assert.Null(cache[key2]);
-
-        Assert.Equal(CacheItemUpdateReason.Expired, updateReason[key1]);
-        Assert.Equal(CacheItemUpdateReason.DependencyChanged, updateReason[key2]);
-    }
-
-    /// <summary>
-    /// This will ensure we can remove the hosting environment after the test is complete since it is a static instance
-    /// </summary>
-    private sealed class HostingEnvironmentContext : IDisposable, IServiceProvider
-    {
-        private readonly Cache _cache;
-
-        private HostingEnvironmentContext(Cache cache)
-        {
-            _cache = cache;
-            HostingEnvironmentAccessor.Current = new(this, Options.Create(new SystemWebAdaptersOptions()));
-        }
-
-        public void Dispose()
-        {
-            HostingEnvironmentAccessor.Current = null;
-        }
-
-        public static HostingEnvironmentContext Initialize(Cache cache)
-        {
-            return new HostingEnvironmentContext(cache);
-        }
-
-        object? IServiceProvider.GetService(Type serviceType)
-        {
-            if (serviceType == typeof(Cache))
-            {
-                return _cache;
-            }
-
-            return null;
-        }
+        //Act via GetService
+        var cacheFromService = contextWrapper.GetService<Cache>();
+        Assert.Same(cache, cacheFromService);
     }
 }
